# improved performance of Memory.get_Empty