@shiftylogic 2017-02-23
This change seems benign. Shouldn't the compiler see that the value returned by the getter is readonly and static and just hardwire it?

Also, all the benchmarks appear to get worse except for plain-text non-pipeline. That seems odd.