@pakrym 2017-02-23
LGTM, I agree with @davidfowl