@KrzysztofCwalina 2017-02-24
@shiftylogic => means a call and the call (to the getter) might have side effects, i.e. does not matter that it is get-only.

{ get; } = OwnerEmptyMemory<T>.Shared.Memory is an initialization syntax (not a call) and will be executed once.